+++ src/mpservice/_common.py
@@ -1,0 +1,5 @@
+import builtins
+
+
+class TimeoutError(builtins.TimeoutError):
+    pass