+++ src/mpservice/mpserver.py
@@ -39,8 +39,7 @@ from typing import Any, Callable, Literal, final
 
 import psutil
 
-from mpservice import TimeoutError
-
+from ._common import TimeoutError
 from ._queues import SingleLane
 from .multiprocessing import MP_SPAWN_CTX, Process, RemoteException
 from .multiprocessing.remote_exception import EnsembleError